+++ src/page/StoreRegistration/StoreRegistration.module.scss
@@ -54,7 +54,7 @@
     padding: 8px 12px;
   }
 
-  &__label-link {
+  &__label-button {
     text-decoration: none;
     display: flex;
     justify-content: center;
@@ -111,7 +111,7 @@
 
     &:checked {
       border: none;
-      background-image: url("../../../assets/svg/storereg/mobile-checked.svg");
+      background-image: url("../../assets/svg/storereg/mobile-checked.svg");
     }
   }
