+++ src/page/StoreRegistration/constant/progress.ts
